requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==23.1.0
    # via quart
aiohttp==3.8.4
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
blinker==1.5
    # via quart
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via quart
colorama==0.4.6
    # via click
cryptg==0.4.0
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
hypercorn==0.14.3
    # via
    #   -r requirements.in
    #   quart
hyperframe==6.0.1
    # via h2
idna==3.4
    # via
    #   requests
    #   yarl
itsdangerous==2.1.2
    # via quart
jinja2==3.1.2
    # via quart
markupsafe==2.1.2
    # via
    #   jinja2
    #   quart
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
priority==2.0.0
    # via hypercorn
pyaes==1.6.1
    # via
    #   pyrogram
    #   telethon
pyasn1==0.5.0
    # via rsa
pyrogram==2.0.106
    # via -r requirements.in
pysocks==1.7.1
    # via pyrogram
python-dotenv==1.0.0
    # via -r requirements.in
quart==0.18.4
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
rsa==4.9
    # via telethon
telethon==1.28.5
    # via -r requirements.in
tgcrypto==1.2.5
    # via -r requirements.in
toml==0.10.2
    # via hypercorn
urllib3==1.26.15
    # via requests
werkzeug==2.2.3
    # via quart
wsproto==1.2.0
    # via hypercorn
yarl==1.9.1
    # via aiohttp
